+++ Tests/SessionReplayTests/SessionReplayModelTests.swift
@@ -706,13 +706,15 @@ class MockSessionReplayModel: SessionReplayModel {
         return .success
     }
     
-    override func captureImage(properties: [String : Any]? = nil) -> Result<Void, CaptureEventError> {
+    override func captureImage(properties: [String : Any]? = nil,
+                               completion: CaptureEventCompletion? = nil) -> Result<Void, CaptureEventError> {
         captureImageCallCount += 1
 
         captureCalled = true
         capturedData = "mock image".data(using: .utf8)
         XCTAssertTrue(Thread.isMainThread, "captureImage should be called on the main thread")
         expectation?.fulfill()
+        completion?(.success(()))
         return .success(())
     }
     
@@ -759,10 +761,12 @@ class MockSessionReplayModel2: SessionReplayModel {
         image: UIImage,
         compressionQuality: CGFloat,
         properties: [String : Any]?,
-        callerIncrementedCounter: Bool
+        callerIncrementedCounter: Bool,
+        completion: CaptureEventCompletion? = nil
     ) {
         let data = "mock image".data(using: .utf8)!
         saveScreenshotToFileSystem(screenshotData: data, properties: properties)
+        completion?(.success(()))
     }
 
     override func saveScreenshotToFileSystem(screenshotData: Data, properties: [String : Any]?) {